Split memory port IO into a separate file
